Moved the 'SharedBuf' to its own module
